Initial renaming and file sort
