+++ src/main/java/com/aws/greengrass/smbridge/Message.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.mqttbridge;
+package com.aws.greengrass.smbridge;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
