+++ src/main/java/com/aws/greengrass/smbridge/auth/CsrGeneratingException.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.mqttbridge.auth;
+package com.aws.greengrass.smbridge.auth;
 
 /**
  * Exception thrown in CSR generation.
