+++ src/main/java/com/aws/greengrass/smbridge/clients/MQTTClient.java
@@ -3,15 +3,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.mqttbridge.clients;
+package com.aws.greengrass.smbridge.clients;
 
 import com.aws.greengrass.componentmanager.KernelConfigResolver;
 import com.aws.greengrass.config.Topics;
 import com.aws.greengrass.logging.api.Logger;
 import com.aws.greengrass.logging.impl.LogManager;
-import com.aws.greengrass.mqttbridge.MQTTBridge;
-import com.aws.greengrass.mqttbridge.Message;
-import com.aws.greengrass.mqttbridge.auth.MQTTClientKeyStore;
+import com.aws.greengrass.smbridge.SMBridge;
+import com.aws.greengrass.smbridge.Message;
+import com.aws.greengrass.smbridge.auth.MQTTClientKeyStore;
 import com.aws.greengrass.util.Coerce;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -99,7 +99,7 @@ public class MQTTClient implements MessageClient {
         this.dataStore = new MemoryPersistence();
         this.serverUri = Coerce.toString(topics.findOrDefault(DEFAULT_BROKER_URI,
                 KernelConfigResolver.PARAMETERS_CONFIG_KEY, BROKER_URI_KEY));
-        this.clientId = Coerce.toString(topics.findOrDefault(MQTTBridge.SERVICE_NAME,
+        this.clientId = Coerce.toString(topics.findOrDefault(SMBridge.SERVICE_NAME,
                 KernelConfigResolver.PARAMETERS_CONFIG_KEY, CLIENT_ID_KEY));
         this.mqttClientKeyStore = mqttClientKeyStore;
         this.mqttClientKeyStore.listenToUpdates(this::reset);
@@ -186,17 +186,6 @@ public class MQTTClient implements MessageClient {
                 LOGGER.atWarn().kv(TOPIC, s).setCause(e).log("Unable to unsubscribe");
             }
         });
-    }
-
-    @Override
-    public void publish(Message message) throws MessageClientException {
-        try {
-            mqttClientInternal
-                    .publish(message.getTopic(), new org.eclipse.paho.client.mqttv3.MqttMessage(message.getPayload()));
-        } catch (MqttException e) {
-            LOGGER.atError().setCause(e).kv(TOPIC, message.getTopic()).log("MQTT Publish failed");
-            throw new MQTTClientException("Failed to publish message", e);
-        }
     }
 
     @Override
