+++ src/main/java/com/aws/greengrass/smbridge/clients/MQTTClientException.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.mqttbridge.clients;
+package com.aws.greengrass.smbridge.clients;
 
 /**
  * Exception thrown by the MQTT Client.
