+++ src/test/java/com/aws/greengrass/smbridge/MessageBridgeTest.java
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.mqttbridge;
+package com.aws.greengrass.smbridge;
 
-import com.aws.greengrass.mqttbridge.clients.MessageClient;
-import com.aws.greengrass.mqttbridge.clients.MessageClientException;
+import com.aws.greengrass.smbridge.clients.MessageClient;
+import com.aws.greengrass.smbridge.clients.MessageClientException;
 import com.aws.greengrass.testcommons.testutilities.GGExtension;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
