+++ src/test/resources/com/aws/greengrass/smbridge/config_with_mapping.yaml
